feat: move docling components to bundle
