+++ src/bundles/docling/pyproject.toml
@@ -1,0 +1,66 @@
+[project]
+name = "lfx-docling"
+version = "0.1.0"
+description = "Docling document processing components as a standalone Langflow Extension Bundle."
+readme = "README.md"
+requires-python = ">=3.10,<3.15"
+license = { text = "MIT" }
+authors = [
+    { name = "Langflow", email = "contact@langflow.org" },
+]
+keywords = ["langflow", "lfx", "extension", "bundle", "docling", "documents"]
+
+# Keep the full local converter and OCR stack optional. docling-core remains a
+# base dependency because the bundle exchanges DoclingDocument objects even when
+# conversion happens remotely.
+dependencies = [
+    "lfx>=0.5.0",
+    "httpx>=0.28.1,<1.0.0",
+    "docling-core>=2.36.1,<3.0.0",
+]
+
+[project.optional-dependencies]
+local = [
+    "docling>=2.36.1,<3.0.0; sys_platform != 'darwin' or platform_machine != 'x86_64'",
+    "tesserocr>=2.8.0",
+    "rapidocr-onnxruntime>=1.4.4",
+    "ocrmac>=1.0.0; sys_platform == 'darwin'",
+    "torch>=2.6.0",
+    "torchvision>=0.21.0",
+]
+chunking = [
+    "lfx-docling[local]",
+    "docling-core[chunking]>=2.36.1,<3.0.0",
+    "tiktoken>=0.7.0",
+]
+image-description = [
+    "lfx-docling[local]",
+    "langchain-docling>=1.1.0",
+]
+all = [
+    "lfx-docling[local,chunking,image-description]",
+]
+
+[project.urls]
+Homepage = "https://github.com/langflow-ai/langflow"
+Documentation = "https://docs.langflow.org/bundles-docling"
+Repository = "https://github.com/langflow-ai/langflow"
+
+[project.entry-points."langflow.extensions"]
+lfx-docling = "lfx_docling"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/lfx_docling"]
+include = ["src/lfx_docling/extension.json", "src/lfx_docling/components/**/*.py"]
+
+[tool.hatch.build.targets.sdist]
+include = [
+    "src/lfx_docling",
+    "extension.json",
+    "README.md",
+    "pyproject.toml",
+]
