+++ src/bundles/docling/src/lfx_docling/__init__.py
@@ -1,0 +1,13 @@
+"""lfx-docling: Docling document processing components."""
+
+from lfx_docling.components.docling.chunk_docling_document import ChunkDoclingDocumentComponent
+from lfx_docling.components.docling.docling_inline import DoclingInlineComponent
+from lfx_docling.components.docling.docling_remote import DoclingRemoteComponent
+from lfx_docling.components.docling.export_docling_document import ExportDoclingDocumentComponent
+
+__all__ = [
+    "ChunkDoclingDocumentComponent",
+    "DoclingInlineComponent",
+    "DoclingRemoteComponent",
+    "ExportDoclingDocumentComponent",
+]
