+++ src/bundles/docling/src/lfx_docling/components/docling/__init__.py
@@ -1,0 +1,11 @@
+from .chunk_docling_document import ChunkDoclingDocumentComponent
+from .docling_inline import DoclingInlineComponent
+from .docling_remote import DoclingRemoteComponent
+from .export_docling_document import ExportDoclingDocumentComponent
+
+__all__ = [
+    "ChunkDoclingDocumentComponent",
+    "DoclingInlineComponent",
+    "DoclingRemoteComponent",
+    "ExportDoclingDocumentComponent",
+]
