+++ src/bundles/docling/src/lfx_docling/extension.json
@@ -1,0 +1,16 @@
+{
+  "$schema": "https://schemas.langflow.org/extension/v1.json",
+  "id": "lfx-docling",
+  "version": "0.1.0",
+  "name": "Docling",
+  "description": "Docling document processing components as a standalone Langflow Extension Bundle.",
+  "lfx": {
+    "compat": ["1"]
+  },
+  "bundles": [
+    {
+      "name": "docling",
+      "path": "components/docling"
+    }
+  ]
+}
