+++ src/lfx/src/lfx/components/__init__.py
@@ -37,7 +37,6 @@ if TYPE_CHECKING:
         datastax,
         deactivated,
         deepseek,
-        docling,
         documentloaders,
         elastic,
         embeddings,
@@ -139,7 +138,6 @@ _dynamic_imports = {
     "datastax": "__module__",
     "deactivated": "__module__",
     "deepseek": "__module__",
-    "docling": "__module__",
     "documentloaders": "__module__",
     "elastic": "__module__",
     "embeddings": "__module__",
@@ -271,7 +269,6 @@ __all__ = [
     "datastax",
     "deactivated",
     "deepseek",
-    "docling",
     "documentloaders",
     "elastic",
     "embeddings",
