+++ src/lfx/src/lfx/extension/migration/migration_table.json
@@ -100,6 +100,86 @@
       "legacy_slot": "ext:ibm:WatsonxEmbeddingsComponent@official-pre-a",
       "target": "ext:ibm:WatsonxEmbeddingsComponent@official",
       "added_in": "1.10.0"
+    },
+    {
+      "bare_class_name": "ChunkDoclingDocumentComponent",
+      "target": "ext:docling:ChunkDoclingDocumentComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "import_path": "lfx.components.docling.chunk_docling_document.ChunkDoclingDocumentComponent",
+      "target": "ext:docling:ChunkDoclingDocumentComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "import_path": "lfx.components.docling.ChunkDoclingDocumentComponent",
+      "target": "ext:docling:ChunkDoclingDocumentComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "legacy_slot": "ext:docling:ChunkDoclingDocumentComponent@official-pre-a",
+      "target": "ext:docling:ChunkDoclingDocumentComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "bare_class_name": "DoclingInlineComponent",
+      "target": "ext:docling:DoclingInlineComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "import_path": "lfx.components.docling.docling_inline.DoclingInlineComponent",
+      "target": "ext:docling:DoclingInlineComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "import_path": "lfx.components.docling.DoclingInlineComponent",
+      "target": "ext:docling:DoclingInlineComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "legacy_slot": "ext:docling:DoclingInlineComponent@official-pre-a",
+      "target": "ext:docling:DoclingInlineComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "bare_class_name": "DoclingRemoteComponent",
+      "target": "ext:docling:DoclingRemoteComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "import_path": "lfx.components.docling.docling_remote.DoclingRemoteComponent",
+      "target": "ext:docling:DoclingRemoteComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "import_path": "lfx.components.docling.DoclingRemoteComponent",
+      "target": "ext:docling:DoclingRemoteComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "legacy_slot": "ext:docling:DoclingRemoteComponent@official-pre-a",
+      "target": "ext:docling:DoclingRemoteComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "bare_class_name": "ExportDoclingDocumentComponent",
+      "target": "ext:docling:ExportDoclingDocumentComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "import_path": "lfx.components.docling.export_docling_document.ExportDoclingDocumentComponent",
+      "target": "ext:docling:ExportDoclingDocumentComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "import_path": "lfx.components.docling.ExportDoclingDocumentComponent",
+      "target": "ext:docling:ExportDoclingDocumentComponent@official",
+      "added_in": "1.10.0"
+    },
+    {
+      "legacy_slot": "ext:docling:ExportDoclingDocumentComponent@official-pre-a",
+      "target": "ext:docling:ExportDoclingDocumentComponent@official",
+      "added_in": "1.10.0"
     }
   ],
   "ambiguous_bare_names": [
